Move helper functions in /controller/texts to utils class and write tests Fixes #1180
