+++ api/Dockerfile
@@ -6,7 +6,7 @@ ENV FASTIFY_ADDRESS=0.0.0.0
 
 WORKDIR /usr/src/app
 COPY ["package.json", "yarn.lock", "./"]
-RUN yarn install --production && mv node_modules ../ && yarn global add fastify-cli@3.1.0 && yarn cache clean
+RUN yarn install --production && mv node_modules ../ && yarn global add fastify-cli && yarn cache clean
 COPY . .
 RUN chown -R node /usr/src/app
 USER node
