+++ api/controllers/texts.js
@@ -4,9 +4,10 @@
 
 'use strict'
 
-const { v4: uuidv4 } = require('uuid')
 const axios = require('axios').default
 const COLLECTIONS = require('../constants/collections.js')
+const { produceVocabularyCollection } = require('../services/utils')
+
 const { ObjectId } = require('mongodb')
 
 async function listTexts(request, reply) {
@@ -60,7 +61,7 @@ async function getText(request, reply) {
   // eslint-disable-next-line putout/putout
   console.log('vocabularyCollection', vocabularyCollection)
 
-  text ? reply.send(text) : reply.notFound('The Text was not found')
+  text ? reply.send(text) : reply.notFound('The text was not found')
 }
 
 async function getTashkeel(request, reply) {
@@ -125,89 +126,6 @@ async function deleteText(request, reply) {
 }
 
 // Helper functions
-
-const shuffleArray = (array) => {
-  return array
-    .map((value) => ({ value, sort: Math.random() }))
-    .sort((a, b) => a.sort - b.sort)
-    .map(({ value }) => value)
-}
-
-function produceVocabularyCollection(text) {
-  const arabicVocabulary = [[]]
-  const englishVocabulary = [[]]
-
-  const maxWordsInBatch = 5
-  let currentBatchNumber = 0
-  let wordsInCurrentBatch = 0
-
-  let numberOfWords = 0
-
-  for (const sentence of text.sentences) {
-    for (const word of sentence.words) {
-      if (word.quiz) {
-        ++numberOfWords
-        // eslint-disable-next-line putout/putout
-        console.log('sentence number:', numberOfWords)
-      }
-    }
-  }
-
-  const maxNumberOfBatches = Math.floor(numberOfWords / maxWordsInBatch)
-
-  for (const sentence of text.sentences) {
-    for (const word of sentence.words) {
-      if (!word.quiz) {
-        // don't add words not suitable for the quiz
-        continue
-      }
-
-      // eslint-disable-next-line putout/putout
-      console.log('max number:', maxNumberOfBatches)
-
-      if (currentBatchNumber === maxNumberOfBatches) {
-        break
-      }
-
-      // Prepare the word to add
-      const wordId = uuidv4()
-
-      const arabicWord = {
-        word: word.arabic,
-        wordId
-      }
-
-      const englishWord = {
-        word: word.english,
-        wordId
-      }
-
-      arabicVocabulary[currentBatchNumber].push(arabicWord)
-      englishVocabulary[currentBatchNumber].push(englishWord)
-
-      ++wordsInCurrentBatch
-
-      // prepare the next batch
-      if (wordsInCurrentBatch === maxWordsInBatch) {
-        // Shuffle words in current batch so they don't appear next each other
-        arabicVocabulary[currentBatchNumber] = shuffleArray(arabicVocabulary[currentBatchNumber])
-        englishVocabulary[currentBatchNumber] = shuffleArray(englishVocabulary[currentBatchNumber])
-        // Proceed to next batch
-        ++currentBatchNumber
-        // Initialize the new batches array
-        arabicVocabulary[currentBatchNumber] = []
-        englishVocabulary[currentBatchNumber] = []
-        wordsInCurrentBatch = 0
-      }
-    }
-  }
-
-  return {
-    numberOfBatches: currentBatchNumber,
-    arabic: arabicVocabulary,
-    english: englishVocabulary
-  }
-}
 
 module.exports = {
   listTexts,
