+++ api/package.json
@@ -18,7 +18,7 @@
     "@fastify/autoload": "^5.4.0",
     "@fastify/cors": "^8.1.0",
     "@fastify/env": "^4.1.0",
-    "@fastify/helmet": "^10.0.0",
+    "@fastify/helmet": "^10.0.1",
     "@fastify/mongodb": "^6.1.0",
     "@fastify/rate-limit": "^7.4.0",
     "@fastify/sensible": "^5.1.1",
@@ -30,7 +30,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-putout": "^16.3.0",
     "eslint-plugin-security": "^1.5.0",
-    "eslint-plugin-unicorn": "^43.0.2",
+    "eslint-plugin-unicorn": "^44.0.0",
     "fastify": "^4.7.0",
     "fastify-cli": "^5.5.0",
     "fastify-healthcheck": "^4.2.0",
